% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{suf_palette}
\alias{suf_palette}
\title{SUffragette- and suffragist-inspired palettes}
\usage{
suf_palette(name, n, type = c("discrete", "continuous"))
}
\arguments{
\item{name}{Name of desired palette. Choices are:
\code{BottleRocket1}, \code{BottleRocket2},  \code{Rushmore1},
\code{london}, \code{oxon},  \code{CarolMan}, \code{hanwell},
\code{chelsea},  \code{classic}}

\item{n}{Number of colors desired. Number of colours available vary by type of palette.
If omitted, uses all colours.}

\item{type}{Either "continuous" or "discrete". Use continuous if you want
to automatically interpolate between colours.
@importFrom graphics rgb rect par image text}
}
\value{
A vector of colours.
}
\description{
SUffragette- and suffragist-inspired palettes
}
\examples{
suf_palette("london")
suf_palette("chelsea", n = 3)

# If you need more colours than normally found in a palette, you
# can use a continuous palette to interpolate between existing
# colours
suf_palette("classic", n = 6, type = "continuous")
}
\keyword{colors}
