% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{suf_palettes}
\alias{suf_palettes}
\title{Complete list of palettes}
\format{An object of class \code{list} of length 6.}
\usage{
suf_palettes
}
\description{
Use \code{\link{suf_palette}} to construct palettes of desired length.
}
\keyword{datasets}
