% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffrage.R
\docType{package}
\name{suffrager}
\alias{suffrager}
\alias{suffrager-package}
\title{suffrager}
\description{
List of palettes from suffragette and suffragist materials
}
\details{
List of palettes from suffragette and suffragist materials
}
